+++ src/main/java/org/itechciv/dashboard/iservice/UploadService.java
@@ -1,7 +1,6 @@
 package org.itechciv.dashboard.iservice;
 
 import java.util.List;
-import java.util.Map;
 
 import org.itechciv.dashboard.helper.CategoryAge;
 import org.springframework.web.multipart.MultipartFile;
@@ -23,7 +22,7 @@ public interface UploadService {
 	List<CategoryAge> getCDCAgeCategory();
 	List<CategoryAge> getNationalAgeCategory(); 
 
-	long getCDCAgeCategorieId(Integer age);
-	long getNationalCategoriesId(Integer age);
+	Integer getCDCAgeCategorieId(Integer age);
+	Integer getNationalCategoriesId(Integer age);
 
 }
